Create new controller for UserAccount and move all methods from userService to userAccountService
