+++ src/main/java/com/school/project/service/UserService.java
@@ -1,7 +1,6 @@
 package com.school.project.service;
 
 import com.school.project.model.entities.User;
-import com.school.project.model.entities.UserAccount;
 
 import java.util.List;
 
@@ -16,9 +15,5 @@ public interface UserService {
     User getUserById(Long id);
 
     List<User> getAllUsers();
-
-    void updateAccountRole(Long userId, Integer statusId);
-
-    UserAccount getUserAccountByUserId(Long id);
 
 }
