+++ src/main/java/com/school/project/service/impl/UserAccountServiceImpl.java
@@ -1,5 +1,6 @@
 package com.school.project.service.impl;
 
+import com.school.project.exception.UserAccountNotFoundException;
 import com.school.project.model.entities.User;
 import com.school.project.model.entities.UserAccount;
 import com.school.project.model.types.UserAccountType;
@@ -7,6 +8,8 @@ import com.school.project.repository.UserAccountRepository;
 import com.school.project.service.UserAccountService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
+import java.util.Optional;
 
 @Service
 public class UserAccountServiceImpl implements UserAccountService {
@@ -23,5 +26,20 @@ public class UserAccountServiceImpl implements UserAccountService {
         userAccountRepository.save(userAccount);
         return userAccount;
     }
+
+    @Override
+    public UserAccount getUserAccountByUserId(Long id) {
+        Optional<UserAccount> account = userAccountRepository.getUserAccountByUserId(id);
+        if (!account.isPresent()) throw new UserAccountNotFoundException("User account with userId : " + id + " is not found");
+        return account.get();
+    }
+
+    @Override
+    public void updateAccountRole(Long userId, Integer statusId){
+        UserAccount account = getUserAccountByUserId(userId);
+        account.setAccountRole(UserAccountType.getTypeById(statusId));
+        userAccountRepository.saveAndFlush(account);
+    }
+
 
 }
